+++ aura-node/src/network.rs
@@ -1,22 +1,22 @@
 use libp2p::{
-    core::upgrade,
-    gossipsub::{self, MessageAuthenticity, Topic},
+    gossipsub::{self, MessageAuthenticity},
     identify,
     kad::{self, store::MemoryStore},
     noise,
-    swarm::{NetworkBehaviour, SwarmBuilder, SwarmEvent},
+    swarm::{NetworkBehaviour, SwarmEvent},
     tcp,
     yamux,
-    PeerId, Transport,
+    PeerId, SwarmBuilder,
 };
 use serde::{Deserialize, Serialize};
+use bincode::{Encode, Decode};
 use std::collections::hash_map::DefaultHasher;
 use std::hash::{Hash, Hasher};
 use std::time::Duration;
-use tracing::{info, warn, error};
+use tracing::{info, warn};
 use crate::config::NetworkConfig;
 
-#[derive(NetworkBehaviour)]
+#[derive(libp2p::swarm::NetworkBehaviour)]
 pub struct AuraNetworkBehaviour {
     gossipsub: gossipsub::Behaviour,
     kademlia: kad::Behaviour<MemoryStore>,
@@ -29,22 +29,22 @@ pub struct NetworkManager {
 }
 
 pub struct NetworkTopics {
-    pub blocks: Topic,
-    pub transactions: Topic,
-    pub did_updates: Topic,
+    pub blocks: gossipsub::IdentTopic,
+    pub transactions: gossipsub::IdentTopic,
+    pub did_updates: gossipsub::IdentTopic,
 }
 
 impl NetworkTopics {
     fn new() -> Self {
         Self {
-            blocks: Topic::new("aura/blocks/1.0.0"),
-            transactions: Topic::new("aura/transactions/1.0.0"),
-            did_updates: Topic::new("aura/did-updates/1.0.0"),
+            blocks: gossipsub::IdentTopic::new("aura/blocks/1.0.0"),
+            transactions: gossipsub::IdentTopic::new("aura/transactions/1.0.0"),
+            did_updates: gossipsub::IdentTopic::new("aura/did-updates/1.0.0"),
         }
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Encode, Decode)]
 pub enum NetworkMessage {
     NewBlock(Vec<u8>),
     NewTransaction(Vec<u8>),
@@ -59,12 +59,7 @@ impl NetworkManager {
         
         info!("Local peer ID: {}", local_peer_id);
         
-        // Create transport
-        let transport = tcp::tokio::Transport::new(tcp::Config::default())
-            .upgrade(upgrade::Version::V1)
-            .authenticate(noise::Config::new(&local_key).unwrap())
-            .multiplex(yamux::Config::default())
-            .boxed();
+        // Transport will be created by SwarmBuilder
         
         // Create Gossipsub
         let message_id_fn = |message: &gossipsub::Message| {
@@ -83,13 +78,16 @@ impl NetworkManager {
         let mut gossipsub = gossipsub::Behaviour::new(
             MessageAuthenticity::Signed(local_key.clone()),
             gossipsub_config,
-        )?;
+        ).map_err(|e| anyhow::anyhow!("Failed to create gossipsub: {}", e))?;
         
         // Create topics and subscribe
         let topics = NetworkTopics::new();
-        gossipsub.subscribe(&topics.blocks)?;
-        gossipsub.subscribe(&topics.transactions)?;
-        gossipsub.subscribe(&topics.did_updates)?;
+        gossipsub.subscribe(&topics.blocks)
+            .map_err(|e| anyhow::anyhow!("Failed to subscribe to blocks topic: {}", e))?;
+        gossipsub.subscribe(&topics.transactions)
+            .map_err(|e| anyhow::anyhow!("Failed to subscribe to transactions topic: {}", e))?;
+        gossipsub.subscribe(&topics.did_updates)
+            .map_err(|e| anyhow::anyhow!("Failed to subscribe to did-updates topic: {}", e))?;
         
         // Create Kademlia
         let kademlia = kad::Behaviour::new(local_peer_id, MemoryStore::new(local_peer_id));
@@ -108,13 +106,20 @@ impl NetworkManager {
         };
         
         // Create swarm
-        let mut swarm = SwarmBuilder::with_tokio_executor(transport, behaviour, local_peer_id).build();
+        let mut swarm = SwarmBuilder::with_existing_identity(local_key.clone())
+            .with_tokio()
+            .with_tcp(tcp::Config::default(), noise::Config::new, yamux::Config::default)
+            .map_err(|e| anyhow::anyhow!("Failed to configure TCP transport: {:?}", e))?
+            .with_behaviour(|_key| Ok(behaviour))
+            .map_err(|e| anyhow::anyhow!("Failed to configure behaviour: {:?}", e))?
+            .build();
         
         // Listen on configured addresses
         for addr in &config.listen_addresses {
             match addr.parse() {
                 Ok(multiaddr) => {
-                    swarm.listen_on(multiaddr)?;
+                    swarm.listen_on(multiaddr)
+                        .map_err(|e| anyhow::anyhow!("Failed to listen on address: {}", e))?;
                     info!("Listening on {}", addr);
                 }
                 Err(e) => {
@@ -125,7 +130,7 @@ impl NetworkManager {
         
         // Connect to bootstrap nodes
         for bootstrap in &config.bootstrap_nodes {
-            if let Ok(multiaddr) = bootstrap.parse() {
+            if let Ok(multiaddr) = bootstrap.parse::<libp2p::Multiaddr>() {
                 match swarm.dial(multiaddr) {
                     Ok(_) => info!("Dialing bootstrap node: {}", bootstrap),
                     Err(e) => warn!("Failed to dial bootstrap node {}: {}", bootstrap, e),
@@ -138,36 +143,42 @@ impl NetworkManager {
     
     pub async fn broadcast_block(&mut self, block_data: Vec<u8>) -> anyhow::Result<()> {
         let message = NetworkMessage::NewBlock(block_data);
-        let data = bincode::serialize(&message)?;
+        let data = bincode::encode_to_vec(&message, bincode::config::standard())
+            .map_err(|e| anyhow::anyhow!("Failed to serialize message: {}", e))?;
         
         self.swarm
             .behaviour_mut()
             .gossipsub
-            .publish(self.topics.blocks.clone(), data)?;
+            .publish(self.topics.blocks.clone(), data)
+            .map_err(|e| anyhow::anyhow!("Failed to publish block: {:?}", e))?;
         
         Ok(())
     }
     
     pub async fn broadcast_transaction(&mut self, tx_data: Vec<u8>) -> anyhow::Result<()> {
         let message = NetworkMessage::NewTransaction(tx_data);
-        let data = bincode::serialize(&message)?;
+        let data = bincode::encode_to_vec(&message, bincode::config::standard())
+            .map_err(|e| anyhow::anyhow!("Failed to serialize message: {}", e))?;
         
         self.swarm
             .behaviour_mut()
             .gossipsub
-            .publish(self.topics.transactions.clone(), data)?;
+            .publish(self.topics.transactions.clone(), data)
+            .map_err(|e| anyhow::anyhow!("Failed to publish transaction: {:?}", e))?;
         
         Ok(())
     }
     
     pub async fn broadcast_did_update(&mut self, did_data: Vec<u8>) -> anyhow::Result<()> {
         let message = NetworkMessage::DidUpdate(did_data);
-        let data = bincode::serialize(&message)?;
+        let data = bincode::encode_to_vec(&message, bincode::config::standard())
+            .map_err(|e| anyhow::anyhow!("Failed to serialize message: {}", e))?;
         
         self.swarm
             .behaviour_mut()
             .gossipsub
-            .publish(self.topics.did_updates.clone(), data)?;
+            .publish(self.topics.did_updates.clone(), data)
+            .map_err(|e| anyhow::anyhow!("Failed to publish DID update: {:?}", e))?;
         
         Ok(())
     }
@@ -181,10 +192,10 @@ impl NetworkManager {
                 Some(SwarmEvent::NewListenAddr { address, .. }) => {
                     info!("Listening on: {}", address);
                 }
-                Some(SwarmEvent::ConnectionEstablished { peer_id, .. }) => {
+                Some(SwarmEvent::ConnectionEstablished { peer_id, connection_id: _, .. }) => {
                     info!("Connected to peer: {}", peer_id);
                 }
-                Some(SwarmEvent::ConnectionClosed { peer_id, .. }) => {
+                Some(SwarmEvent::ConnectionClosed { peer_id, connection_id: _, .. }) => {
                     info!("Disconnected from peer: {}", peer_id);
                 }
                 _ => {}
@@ -195,19 +206,20 @@ impl NetworkManager {
     async fn handle_behaviour_event(&mut self, event: AuraNetworkBehaviourEvent) {
         match event {
             AuraNetworkBehaviourEvent::Gossipsub(gossipsub::Event::Message {
-                propagation_source,
-                message_id,
+                propagation_source: _,
+                message_id: _,
                 message,
             }) => {
                 let topic = message.topic.clone();
                 
-                match bincode::deserialize::<NetworkMessage>(&message.data) {
+                match bincode::decode_from_slice::<NetworkMessage, _>(&message.data, bincode::config::standard())
+                    .map(|(msg, _)| msg) {
                     Ok(network_msg) => {
-                        if topic == self.topics.blocks {
+                        if topic == self.topics.blocks.hash() {
                             self.handle_new_block(network_msg).await;
-                        } else if topic == self.topics.transactions {
+                        } else if topic == self.topics.transactions.hash() {
                             self.handle_new_transaction(network_msg).await;
-                        } else if topic == self.topics.did_updates {
+                        } else if topic == self.topics.did_updates.hash() {
                             self.handle_did_update(network_msg).await;
                         }
                     }
@@ -216,7 +228,7 @@ impl NetworkManager {
                     }
                 }
             }
-            AuraNetworkBehaviourEvent::Identify(identify::Event::Received { peer_id, info }) => {
+            AuraNetworkBehaviourEvent::Identify(identify::Event::Received { peer_id, info, connection_id: _ }) => {
                 info!("Identified peer {}: {:?}", peer_id, info.protocol_version);
             }
             _ => {}
@@ -246,4 +258,3 @@ impl NetworkManager {
 }
 
 use futures::StreamExt;
-use libp2p::swarm::NetworkBehaviour as NetworkBehaviourTrait;